# bindings/nodejs/modules/open-intent-rest-chatbot/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: open-intent REST API
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /model:
    x-swagger-router-controller: rest-api
    get:
      description: Returns the current model of the chatbot
      operationId: getmodel
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetModelResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Set the model of the chatbot
      operationId: setmodel
      parameters:
        - name: botmodel
          in: body
          schema:
            $ref: "#/definitions/ModelObject"
      responses:
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/StatusResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /state/{sessionId}:
    x-swagger-router-controller: rest-api
    get:
      description: Returns the current state of a user session
      operationId: getstate
      parameters:
        - name: sessionId
          in: path
          description: The user session ID
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetStateResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Set the current state of a user session
      operationId: setstate
      consumes:
      - application/x-www-form-urlencoded
      produces:
      - application/json
      parameters:
        - name: sessionId
          in: path
          description: The user session ID
          required: true
          type: string
        - name: state
          in: formData
          description: The state to set
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/StatusResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /talk/{sessionId}:
    x-swagger-router-controller: rest-api
    post:
      description: Talk to the chatbot, call the user commands and get the replies.
      operationId: talk
      consumes:
      - application/x-www-form-urlencoded
      parameters:
        - name: sessionId
          in: path
          description: The user session ID
          required: true
          type: string
        - name: message
          in: formData
          description: The message to send to the bot
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TalkResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"


  /swagger:
    x-swagger-pipe: swagger_raw



definitions:
  ModelObject:
    properties:
      commands:
        type: object
      model:
        type: object

  GetStateResponse:
    required:
      - state
    properties:
      state:
        type: string

  GetModelResponse:
    required:
      - model
    properties:
      model:
        type: object

  TalkResponse:
      required:
        - replies
      properties:
        replies:
          type: array
          items:
            type: string

  StatusResponse:
    properties:
      message:
        type: string

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string